video: tegra: host: Use wall clock time in ms
When waiting for an event with a timeout, use wall time for checking
if timeout has been exceeded. Also changes timeouts to be managed as
milliseconds in unsigned long to unify with the rest of driver.

Changes infinite timeouts to sane maximums.

Bug 1342089

Reviewed-on: http://git-master/r/265939
Tested-by: Juha Tukkinen <jtukkinen@nvidia.com>

+++ drivers/video/tegra/host/gk20a/gk20a.h
@@ -94,7 +94,7 @@ struct gk20a {
 	void (*remove_support)(struct platform_device *);
 };
 
-static inline u32 gk20a_get_gr_idle_timeout(struct gk20a *g)
+static inline unsigned long gk20a_get_gr_idle_timeout(struct gk20a *g)
 {
 	return g->timeouts_enabled ?
 		g->gr_idle_timeout_default : MAX_SCHEDULE_TIMEOUT;
